+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## 6-Jun-20
+
+- Deployment to GitHub Pages
+
 ## 5-Jun-20
 
 - VuePress set up for 100 material